Revert "Revert "Revert "Merge branch 'feature/JNG-2630_Col_value_calculation'"""
This reverts commit 693fc8753be56f3bba05d9507f75de1046d2f41c.

+++ lib/src/judo_button.dart
@@ -19,7 +19,7 @@ class JudoButton extends StatelessWidget {
     this.alignment = Alignment.centerLeft,
   }) : super(key: key);
 
-  final double col;
+  final int col;
   final double row;
   final EdgeInsets padding;
   final String label;
