+++ lib/src/judo_date_time_input.dart
@@ -21,7 +21,7 @@ class JudoDateTimeInput extends StatefulWidget {
     this.alignment = Alignment.centerLeft,
   }) : super(key: key);
 
-  final double col;
+  final int col;
   final double row;
   final bool mandatory;
   final String label;
